requirements.txt
langchain==0.3.27
langchain_chroma==0.2.6
langchain_community==0.3.30
langchain_groq==0.3.8
langchain_huggingface==0.3.1
python-dotenv==1.1.1
streamlit==1.48.1
